requirements.txt
streamlit
langchain
openai
faiss-cpu
python-dotenv
PyPDF2
sentence-transformers
huggingface_hub
nltk

